feat: Refactor document synchronization logic; remove heavy requirements, add doc extractor controller and repository
