+++ src/controllers/doc_chat_controller.py
@@ -43,29 +43,6 @@ class DocChatController:
             self.log.error("Upload document failed", error_msg=error_msg)
             raise HTTPException(status_code=500, detail=f"Unexpected error during document upload.")
         
-    async def sync_documents(self, client_id: str, product_id: str):
-        """
-        Endpoint to synchronize documents for chat analysis.
-        """
-        try:
-            self.log.info("Document synchronization started")
-
-            await self.repository.sync_documents(client_id=client_id, product_id=product_id)
-
-            self.log.info("Document synchronization completed successfully")
-
-            return JSONResponse(content="Document synchronization completed successfully")
-
-        except ValueError as ve:
-            error_msg = CustomException(str(ve), sys).__str__()
-            self.log.error("Sync documents failed", error_msg=error_msg)
-            raise HTTPException(status_code=400, detail=f"{ve}")
-
-        except Exception as e:
-            error_msg = CustomException(str(e), sys).__str__()
-            self.log.error("Sync documents failed", error_msg=error_msg)
-            raise HTTPException(status_code=500, detail=f"Unexpected error during document synchronization.")
-        
     async def vectorize_document(self, client_id: str, product_id: str):
         """
         Endpoint to vectorize a document for chat analysis.
@@ -119,11 +96,6 @@ endpoints = [
         "path": "/upload",
         "method": "post",
         "handler": doc_chat_controller.upload_document
-    },
-    {
-        "path": "/sync",
-        "method": "post",
-        "handler": doc_chat_controller.sync_documents
     },
     {
         "path": "/vectorize",
